requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.2
django-allauth==0.49.0
django-cache-url==3.3.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rosetta==0.9.8
environs==9.5.0
idna==3.3
jdatetime==4.1.0
marshmallow==3.14.1
oauthlib==3.2.0
Pillow==9.1.1
polib==1.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
python-decouple
openai>=0.27.0


